Get collections

Get collections

get_collections(connection = NULL, db_version = "current", db_args = NULL)

Arguments

connection
A connection to the CHILDES database

db_version
String of the name of database version to use

db_args
List with host, user, and password defined

Value

A `tbl` of Collection data. If `connection` is supplied, the result remains a remote query, otherwise it is retrieved into a local tibble.

Examples

# \donttest{
get_collections()

#> Using current database version: '2018.1'.

#> # A tibble: 20 x 2
#>    collection_id collection_name
#>            <int> <chr>          
#>  1             1 Spanish        
#>  2             2 French         
#>  3             3 Eng-NA         
#>  4             4 Clinical-MOR   
#>  5             5 Biling         
#>  6             6 German         
#>  7             7 Celtic         
#>  8             8 XLing          
#>  9             9 Chinese        
#> 10            10 DutchAfrikaans 
#> 11            11 Frogs          
#> 12            12 Dutch          
#> 13            13 Eng-UK         
#> 14            14 Slavic         
#> 15            15 Romance        
#> 16            16 Japanese       
#> 17            17 Scandinavian   
#> 18            18 Clinical       
#> 19            19 EastAsian      
#> 20            20 Other          

# }