Get the utterances surrounding a token(s)

Get the utterances surrounding a token(s)

get_contexts(collection = NULL, language = NULL, corpus = NULL,
  role = NULL, role_exclude = NULL, age = NULL, sex = NULL,
  target_child = NULL, token, window = c(0, 0),
  remove_duplicates = TRUE, connection = NULL,
  db_version = "current", db_args = NULL)

Arguments

collection
A character vector of one or more names of collections

language
A character vector of one or more languages

corpus
A character vector of one or more names of corpora

role
A character vector of one or more roles to include

role_exclude
A character vector of one or more roles to exclude

age
A numeric vector of an age or a min age (inclusive) and max age (exclusive) in months

sex
A character vector of values "male" and/or "female"

target_child
A character vector of one or more names of children

token
A character vector of one or more token patterns (`%` matches any number of wildcard characters, `_` matches exactly one wildcard character)

window
A length 2 numeric vector of how many utterances before and after each utterance containing the target token to retrieve

remove_duplicates
A boolean indicating whether to remove duplicate utterances from the results

connection
A connection to the CHILDES database

db_version
String of the name of database version to use

db_args
List with host, user, and password defined

Value

A 'tbl' of Utterance data, filtered down by supplied arguments.

Examples

get_contexts(target_child = "Shem", token = "dog")

#> Using current database version: '2018.1'.

#> Getting data from 1 child in 1 corpus ...

#> Error in .local(conn, statement, ...): could not run statement: Operand should contain 1 column(s)