Get types

Get types

get_types(
  collection = NULL,
  language = NULL,
  corpus = NULL,
  role = NULL,
  role_exclude = NULL,
  age = NULL,
  sex = NULL,
  target_child = NULL,
  type = NULL,
  connection = NULL,
  db_version = "current",
  db_args = NULL
)

Arguments

collection
A character vector of one or more names of collections

language
A character vector of one or more languages

corpus
A character vector of one or more names of corpora

role
A character vector of one or more roles to include

role_exclude
A character vector of one or more roles to exclude

age
A numeric vector of an single age value or a min age value and max age value (inclusive) in months. For a single age value, participants are returned for which that age is within their age range; for two ages, participants are returned for whose age overlaps with the interval between those two ages.

sex
A character vector of values "male" and/or "female"

target_child
A character vector of one or more names of children

type
A character vector of one or more type patterns (` number of wildcard characters, `_` matches exactly one wildcard character)

connection
A connection to the CHILDES database

db_version
String of the name of database version to use

db_args
List with host, user, and password defined

Value

A `tbl` of Type data, filtered down by supplied arguments. If `connection` is supplied, the result remains a remote query, otherwise it is retrieved into a local tibble.

Examples

# \donttest{
get_types()

#> Using current database version: '2020.1'.

#> Getting data from 7085 children in 375 corpora...

#> # A tibble: 17,897,045 x 14
#>       id gloss count speaker_role language target_child_na… target_child_age
#>    <int> <chr> <int> <chr>        <chr>    <chr>                       <dbl>
#>  1     1 mhm       1 Target_Child eng      Nathan                       78.0
#>  2     2 oh        1 Target_Child eng      Nathan                       78.0
#>  3     3 I've      1 Target_Child eng      Nathan                       78.0
#>  4     4 seen      1 Target_Child eng      Nathan                       78.0
#>  5     5 a         3 Target_Child eng      Nathan                       78.0
#>  6     6 movie     1 Target_Child eng      Nathan                       78.0
#>  7     7 of        2 Target_Child eng      Nathan                       78.0
#>  8     8 that      1 Target_Child eng      Nathan                       78.0
#>  9     9 yeah      3 Target_Child eng      Nathan                       78.0
#> 10    10 one       1 Target_Child eng      Nathan                       78.0
#> # … with 17,897,035 more rows, and 7 more variables: target_child_sex <chr>,
#> #   collection_name <chr>, collection_id <int>, corpus_id <int>,
#> #   speaker_id <int>, target_child_id <int>, transcript_id <int>

# }