Get utterances

Get utterances

get_utterances(
  collection = NULL,
  language = NULL,
  corpus = NULL,
  role = NULL,
  role_exclude = NULL,
  age = NULL,
  sex = NULL,
  target_child = NULL,
  connection = NULL,
  db_version = "current",
  db_args = NULL
)

Arguments

collection
A character vector of one or more names of collections

language
A character vector of one or more languages

corpus
A character vector of one or more names of corpora

role
A character vector of one or more roles to include

role_exclude
A character vector of one or more roles to exclude

age
A numeric vector of an single age value or a min age value (inclusive) and max age value (exclusive) in months. For a single age value, participants are returned for which that age is within their age range; for two ages, participants are returned for whose age overlaps with the interval between those two ages.

sex
A character vector of values "male" and/or "female"

target_child
A character vector of one or more names of children

connection
A connection to the CHILDES database

db_version
String of the name of database version to use

db_args
List with host, user, and password defined

Value

A `tbl` of Utterance data, filtered down by supplied arguments. If `connection` is supplied, the result remains a remote query, otherwise it is retrieved into a local tibble.

Examples

# \donttest{
get_utterances(target_child = "Shem")

#> Using current database version: '2018.1'.

#> Getting data from 1 child in 1 corpus ...

#> # A tibble: 42,467 x 25
#>        id speaker_id utterance_order transcript_id corpus_id gloss num_tokens
#>     <int>      <int>           <int>         <int>     <int> <chr>      <int>
#>  1 776196       2454               1          2765        29 yeah           1
#>  2 776202       2455               2          2765        29 what…          5
#>  3 776217       2454               3          2765        29 ow             1
#>  4 776227       2455               4          2765        29 ow             1
#>  5 776235       2454               5          2765        29 yeah           1
#>  6 776244       2455               6          2765        29 want…          6
#>  7 776261       2454               7          2765        29 ow             1
#>  8 776267       2455               8          2765        29 ow w…          3
#>  9 776279       2455               9          2765        29 Rich…          1
#> 10 776284       2454              10          2765        29 no             1
#> # … with 42,457 more rows, and 18 more variables: stem <chr>,
#> #   part_of_speech <chr>, speaker_code <chr>, speaker_name <chr>,
#> #   speaker_role <chr>, target_child_id <int>, target_child_age <dbl>,
#> #   target_child_name <chr>, target_child_sex <chr>, type <chr>,
#> #   media_end <dbl>, media_start <dbl>, media_unit <chr>, collection_id <int>,
#> #   collection_name <chr>, num_morphemes <int>, language <chr>,
#> #   corpus_name <chr>

# }